Fix remaining clippy warnings across all crates
- Fix type mismatch in cam-binder metadata.rs (value_offset is u32)
- Add #[allow(unused_imports)] for conditional imports in cam-onnx
- Add Safety docs to unsafe functions in cam-hal-android
- Remove unnecessary casts (u16, isize) in cam-hal-android
- Use or_default() instead of or_insert_with(Vec::new)
- Simplify identical if-else blocks
- Add v4l2 feature flag to cam-hal-android Cargo.toml
- Remove unnecessary map_err identity functions in cam-binder/aidl.rs

All tests pass, zero clippy warnings across workspace.

+++ cam-rust/cam-hal-android/src/lib.rs
@@ -364,7 +364,7 @@ unsafe extern "C" fn device_configure_streams(
     let adapter = &mut *adapter_ptr;
 
     for i in 0..cfg.num_streams as isize {
-        let stream = *cfg.streams.offset(i as isize);
+        let stream = *cfg.streams.offset(i);
         let s = &mut *stream;
         log::debug!(
             "  stream[{}]: {}x{} fmt=0x{:x}",
